Update weather model and Clean up code
