+++ data/src/main/java/com/nuweather/data/WeatherRepositoryImpl.kt
@@ -1,19 +1,19 @@
 package com.nuweather.data
 
-import com.nuweather.data.model.CurrentWeatherMapper
+import com.nuweather.data.mapper.WeatherMapper
 import com.nuweather.data.remote.api.WeatherApi
-import com.nuweather.domain.model.CurrentWeather
+import com.nuweather.domain.model.Weather
 import com.nuweather.domain.repository.WeatherRepository
 import io.reactivex.Single
 
 class WeatherRepositoryImpl constructor(
     private val weatherApi: WeatherApi,
-    private val currentWeatherMapper: CurrentWeatherMapper
+    private val weatherMapper: WeatherMapper
 ) : WeatherRepository {
 
-    override fun getCurrentWeather(query: String): Single<CurrentWeather> {
+    override fun getCurrentWeather(query: String): Single<Weather> {
         return weatherApi.getCurrentWeather(query).map {
-            currentWeatherMapper.mapToDomain(it.mapToEntity())
+            weatherMapper.mapToDomain(it.mapToEntity())
         }
     }
 }
