+++ data/src/main/java/com/nuweather/data/model/WindEntity.kt
@@ -1,0 +1,8 @@
+package com.nuweather.data.model
+
+import com.nuweather.data.base.Entity
+
+data class WindEntity(
+    val speed: Double,
+    val deg: Double
+) : Entity()
