+++ domain/src/main/java/com/nuweather/domain/model/Sys.kt
@@ -1,0 +1,10 @@
+package com.nuweather.domain.model
+
+data class Sys(
+    val type: Int,
+    val id: Long,
+    val message: Double,
+    val country: String,
+    val sunrise: Long,
+    val sunset: Long
+) : Model()
