+++ domain/src/main/java/com/nuweather/domain/model/Wind.kt
@@ -1,0 +1,6 @@
+package com.nuweather.domain.model
+
+data class Wind(
+    val speed: Double,
+    val deg: Double
+) : Model()
