+++ domain/src/main/java/com/nuweather/domain/repository/WeatherRepository.kt
@@ -1,8 +1,8 @@
 package com.nuweather.domain.repository
 
-import com.nuweather.domain.model.CurrentWeather
+import com.nuweather.domain.model.Weather
 import io.reactivex.Single
 
 interface WeatherRepository : Repository {
-    fun getCurrentWeather(query: String): Single<CurrentWeather>
+    fun getCurrentWeather(query: String): Single<Weather>
 }
