+++ presentation/src/main/java/com/nuweather/model/WindItem.kt
@@ -1,0 +1,8 @@
+package com.nuweather.model
+
+import com.nuweather.base.ModelItem
+
+data class WindItem(
+    val speed: Double,
+    val deg: Double
+) : ModelItem()
