+++ presentation/src/main/res/layout/fragment_current_weather.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:bind="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     tools:context=".feature.MainActivity">
 
@@ -24,7 +25,7 @@
             android:layout_width="@dimen/dp_0"
             android:layout_height="wrap_content"
             android:layout_marginStart="@dimen/dp_16"
-            android:layout_marginTop="@dimen/dp_24"
+            android:layout_marginTop="@dimen/dp_28"
             android:layout_marginEnd="@dimen/dp_8"
             android:fontFamily="sans-serif-light"
             android:text="@{viewModel.currentWeatherItem.city}"
@@ -155,13 +156,13 @@
 
         <android.support.v7.widget.AppCompatImageView
             android:id="@+id/weatherImage"
-            imageResource="@{viewModel.weatherImage}"
             android:layout_width="@dimen/dp_128"
             android:layout_height="@dimen/dp_128"
             android:layout_marginStart="@dimen/dp_16"
-            android:layout_marginTop="@dimen/dp_24"
+            android:layout_marginTop="@dimen/dp_16"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/addressText"
+            bind:imageResource="@{viewModel.weatherImage}"
             tools:srcCompat="@drawable/ic_cloud_lightning_sun" />
 
         <TextView
