+++ presentation/src/main/res/layout/fragment_home.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:bind="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     tools:context=".feature.MainActivity">
 
@@ -18,7 +19,6 @@
 
         <android.support.v7.widget.AppCompatImageView
             android:id="@+id/weatherBackground"
-            imageResource="@{viewModel.weatherBackground}"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:scaleType="centerCrop"
@@ -26,7 +26,8 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toTopOf="parent"
+            bind:imageResource="@{viewModel.weatherBackground}" />
 
         <FrameLayout
             android:id="@+id/currentWeather"
