+++ presentation/src/main/res/layout/progress_dialog.xml
@@ -2,6 +2,4 @@
 <ProgressBar xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/progressBar"
     android:layout_width="wrap_content"
-    android:layout_height="wrap_content">
-
-</ProgressBar>
+    android:layout_height="wrap_content" />
